Update
updated more parts of the home page.
Interesting to learn about the formats, i think i really took in more about the margins and how the differences between padding, margins and width work in relation to each other.
Still confused though so i'll probably have to watch a video to help explain it.