Update ~1st Stage Complete
I finished the whole homepage template. Now i have to go back and finish my customisations for the website.

Make sure the text is the right size, colour; images are in the right place, look good and update some of the information.

After that i'll move onto the next part: Next HTML Files!