Update: Base Complete
Completed the basic tutorial.

I'm not paying for it to be uploaded online, i'll be using GitPages to show my work instead as it's free and uploads automatically. Works out a lot better as it's just my silly lil project.

I have a function for php to email concerns to me but I doubt it'll actually function in GitPages, it just looks cool that it *could* work in future.