Comment
A small issue when hosting the website through GitPages.

Apparently GitPages runs on Linux, which cares about case sensitivity. Since i like JS i've been using the syntax in my file names as a joke but now it's biting me in the ass and not loading the images in my "/imageBank/merchModels" because of that.

So we're going to fix that, like it says.

I can't wait.

I just can't wait for this.

It's going to be so much fun.

Honestly the highlight of my day.

What could be better than going over all the code you've worked on for the past 2 weeks+ and making sure each file is lowercase.

The dream really.