Update: About Page
I updated the header to have a line above it which helps it to appear separated and also look interesting and interactive to the user as it bounces in and out.

I also finished my Achievements section which I made from scratch using what I learnt and some tips from fellow, more experienced devs. This includes changing the font colour and spacing around.

My next steps for these sections will  be to check that it's formatted adequately for mobile devices.

Oh, another slight adjustment I made was to the opacity in the index model images. Where the hover description used to be slightly tinted as you hovered over it, it's not almost completely opaque for better readability and a cleaner look.