Update: Clothes Section - in-process
Currently working on what I want this page to be, I have some comments in the file which are helping me image what I want and i've been looking at different merch websites for references and ideas.

So far I have made a cool box.

I want to make a transition for the clothes so it will scroll through by default and it will also let you scroll through the clothes.