Update: Slider
Trying to create an interactive slider which people can use to skim through the products we've sold as standard in the past.

I want to make interactive buttons so people don't have to use the bars at the bottom; but I also want it to prompt them or transition on it's own.

After I figure that out i'll be doing it for some other sections on that page to showcase models and fan subs featuring the merch.